+++ src/domain/auth/auth.module.ts
@@ -8,9 +8,8 @@ import { UserRepository } from '#/infrastructure/drizzle/repository/user.reposit
 import { AuthController } from './auth.controller';
 import { AdminGuard } from './guard/admin.guard';
 import { AuthenticationGuard } from './guard/auth.guard';
-import { KakaoAuthGuard } from './guard/kakao-auth.guard';
 import { AuthTokenService } from './service/auth-token.service';
-import { SocialAuthService } from './service/social-auth.service';
+import { KakaoAuthService } from './service/kakao-auth.service';
 
 @Module({
 	imports: [
@@ -23,11 +22,10 @@ import { SocialAuthService } from './service/social-auth.service';
 	providers: [
 		// Service
 		AuthTokenService,
-		SocialAuthService,
+		KakaoAuthService,
 		// Guard
 		AuthenticationGuard,
 		AdminGuard,
-		KakaoAuthGuard,
 		// Config
 		JwtConfig,
 		// Repository
